pyproject.toml
[tool.poetry]
name = "algorithmic-options-trading"
version = "0.1.0"
description = ""
authors = ["<mutgan7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
setuptools = "72.0"
kiteconnect = "^5.0.1"
python-dotenv = "^1.0.1"
redis = "^5.1.1"
fastapi = { extras = ["standard"], version = "^0.115.3" }
requests = "^2.32.3"
pandas = "^2.2.3"
redis-om = "^0.3.2"
jupyter = "^1.1.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
